backend/requirements.txt
fastapi
uvicorn
fpdf2
pillow
python-multipart
